starting shaman
