+++ sim/shaman/item_sets_pve.go
@@ -9,129 +9,16 @@ import (
 	"github.com/wowsims/classic/sim/core/stats"
 )
 
-///////////////////////////////////////////////////////////////////////////
-//                            SoD Phase 2 Item Sets
-///////////////////////////////////////////////////////////////////////////
-
-var ItemSetElectromanticStormbringer = core.NewItemSet(core.ItemSet{
-	Name: "Electromantic Stormbringer's Chain",
-	Bonuses: map[int32]core.ApplyEffect{
-		2: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.SpellPower, 12)
-		},
-		3: func(agent core.Agent) {
-			shaman := agent.(ShamanAgent).GetShaman()
-			shaman.OnSpellRegistered(func(spell *core.Spell) {
-				if spell.SpellCode == SpellCode_ShamanLightningBolt {
-					spell.DefaultCast.CastTime -= time.Millisecond * 100
-				}
-			})
-		},
-	},
-})
-
-///////////////////////////////////////////////////////////////////////////
-//                            SoD Phase 3 Item Sets
-///////////////////////////////////////////////////////////////////////////
-
-var OstracizedBerserksBattlemail = core.NewItemSet(core.ItemSet{
-	Name: "Ostracized Berserker's Battlemail",
-	Bonuses: map[int32]core.ApplyEffect{
-		2: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.AttackPower, 20)
-		},
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-
-			procAura := c.GetOrRegisterAura(core.Aura{
-				Label:     "Fiery Strength Proc",
-				ActionID:  core.ActionID{SpellID: 449932},
-				Duration:  time.Second * 12,
-				MaxStacks: 10,
-				OnStacksChange: func(aura *core.Aura, sim *core.Simulation, oldStacks, newStacks int32) {
-					statsDelta := float64(newStacks-oldStacks) * 5.0
-					aura.Unit.AddStatDynamic(sim, stats.AttackPower, statsDelta)
-				},
-			})
-
-			core.MakeProcTriggerAura(&c.Unit, core.ProcTrigger{
-				Name:     "Fiery Strength",
-				Callback: core.CallbackOnSpellHitDealt | core.CallbackOnPeriodicDamageDealt,
-				Outcome:  core.OutcomeLanded,
-				ProcMask: core.ProcMaskDirect,
-				Handler: func(sim *core.Simulation, spell *core.Spell, result *core.SpellResult) {
-					if spell.SpellSchool.Matches(core.SpellSchoolFire) {
-						procAura.Activate(sim)
-						procAura.AddStack(sim)
-					}
-				},
-			})
-		},
-	},
-})
-
-var ItemSetEmeraldChainmail = core.NewItemSet(core.ItemSet{
-	Name: "Emerald Chainmail",
-	Bonuses: map[int32]core.ApplyEffect{
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.Stamina, 10)
-		},
-		6: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.SpellPower, 12)
-		},
-	},
-})
-
-var ItemSetEmeraldScalemail = core.NewItemSet(core.ItemSet{
-	Name: "Emerald Scalemail",
-	Bonuses: map[int32]core.ApplyEffect{
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.Stamina, 10)
-		},
-		6: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.AttackPower, 20)
-		},
-	},
-})
-
-var ItemSetEmeraldLadenChain = core.NewItemSet(core.ItemSet{
-	Name: "Emerald Laden Chain",
-	Bonuses: map[int32]core.ApplyEffect{
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.Stamina, 10)
-		},
-		6: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.HealingPower, 22)
-		},
-	},
-})
-
-///////////////////////////////////////////////////////////////////////////
-//                            SoD Phase 4 Item Sets
-///////////////////////////////////////////////////////////////////////////
-
 var ItemSetTheFiveThunders = core.NewItemSet(core.ItemSet{
 	Name: "The Five Thunders",
 	Bonuses: map[int32]core.ApplyEffect{
-		// +40 Attack Power, up to 23 increased damage from spells, and up to 44 increased healing from spells.
+		// +8 All Resistances.
 		2: func(agent core.Agent) {
 			c := agent.GetCharacter()
-			c.AddStats(stats.Stats{
-				stats.AttackPower:       40,
-				stats.RangedAttackPower: 40,
-				stats.SpellDamage:       23,
-				stats.HealingPower:      44,
-			})
+			c.AddResistances(8)
 		},
-		// 6% chance on mainhand autoattack and 4% chance on spellcast to increase your damage and healing done by magical spells and effects by up to 95 for 10 sec.
+		// Chance on spell cast to increase your damage and healing by up to 95 for 10 sec.
+		// (Proc chance: 4%)
 		4: func(agent core.Agent) {
 			c := agent.GetCharacter()
 
@@ -141,14 +28,6 @@ var ItemSetTheFiveThunders = core.NewItemSet(core.ItemSet{
 			}
 
 			core.MakeProcTriggerAura(&c.Unit, core.ProcTrigger{
-				Name:       "Item - The Furious Storm Proc (MH Auto)",
-				Callback:   core.CallbackOnSpellHitDealt,
-				Outcome:    core.OutcomeLanded,
-				ProcMask:   core.ProcMaskMeleeMHAuto,
-				ProcChance: 0.06,
-				Handler:    handler,
-			})
-			core.MakeProcTriggerAura(&c.Unit, core.ProcTrigger{
 				Name:       "Item - The Furious Storm Proc (Spell Cast)",
 				Callback:   core.CallbackOnCastComplete,
 				ProcMask:   core.ProcMaskSpellDamage | core.ProcMaskSpellHealing,
@@ -156,7 +35,7 @@ var ItemSetTheFiveThunders = core.NewItemSet(core.ItemSet{
 				Handler:    handler,
 			})
 		},
-		// +8 All Resistances.
+		// Increases damage and healing done by magical spells and effects by up to 23.
 		6: func(agent core.Agent) {
 			c := agent.GetCharacter()
 			c.AddResistances(8)
@@ -169,198 +48,39 @@ var ItemSetTheFiveThunders = core.NewItemSet(core.ItemSet{
 	},
 })
 
-var ItemSetEarthfuryEruption = core.NewItemSet(core.ItemSet{
-	Name: "Earthfury Eruption",
+var ItemSetTheEarthfury = core.NewItemSet(core.ItemSet{
+	Name: "The Earthfury",
 	Bonuses: map[int32]core.ApplyEffect{
-		// The radius of your totems that affect friendly targets is increased to 40 yd.
+		// The radius of your totems that affect friendly targets is increased to 30 yd.
 		2: func(agent core.Agent) {
 			// Nothing to do
 		},
-		// Your Lightning Bolt critical strikes have a 35% chance to reset the cooldown on Lava Burst and Chain Lightning and make the next Lava Burst, Chain Heal, or Chain Lightning within 10 sec instant.
+		// After casting your Healing Wave or Lesser Healing Wave spell, gives you a 25% chance to gain Mana equal to 35% of the base cost of the spell.
 		4: func(agent core.Agent) {
-			shaman := agent.(ShamanAgent).GetShaman()
-			shaman.GetOrRegisterAura(core.Aura{
-				Label:    "S03 - Item - T1 - Shaman - Elemental 4P Bonus",
-				Duration: core.NeverExpires,
-				OnReset: func(aura *core.Aura, sim *core.Simulation) {
-					aura.Activate(sim)
-				},
-				OnSpellHitDealt: func(aura *core.Aura, sim *core.Simulation, spell *core.Spell, result *core.SpellResult) {
-					if spell.SpellCode == SpellCode_ShamanLightningBolt && spell.ProcMask.Matches(core.ProcMaskSpellDamage) && result.DidCrit() && sim.Proc(.35, "Power Surge") {
-						shaman.PowerSurgeDamageAura.Activate(sim)
-					}
-				},
-			})
+			// Nothing to do
 		},
-		// Lava Burst now also refreshes the duration of Flame Shock on your target back to 12 sec.
+		// Your Healing Wave will now jump to additional nearby targets. Each jump reduces the effectiveness of the heal by 80%, and the spell will jump to up to two additional targets.
 		6: func(agent core.Agent) {
-			shaman := agent.(ShamanAgent).GetShaman()
-			core.MakePermanent(shaman.GetOrRegisterAura(core.Aura{
-				Label: "S03 - Item - T1 - Shaman - Elemental 6P Bonus",
-				OnCastComplete: func(aura *core.Aura, sim *core.Simulation, spell *core.Spell) {
-					if spell.SpellCode == SpellCode_ShamanLavaBurst {
-						for _, spell := range shaman.FlameShock {
-							if spell == nil {
-								continue
-							}
-
-							if dot := spell.Dot(shaman.CurrentTarget); dot.IsActive() {
-								dot.NumberOfTicks = dot.OriginalNumberOfTicks
-								dot.Rollover(sim)
-							}
-						}
-					}
-				},
-			}))
+			// Nothing to do
 		},
 	},
 })
 
-var ItemSetEarthfuryRelief = core.NewItemSet(core.ItemSet{
-	Name: "Earthfury Relief",
+var ItemSetTheTenStorms = core.NewItemSet(core.ItemSet{
+	Name: "The Ten Storms",
 	Bonuses: map[int32]core.ApplyEffect{
-		// The radius of your totems that affect friendly targets is increased to 40 yd.
+		// Increases the amount healed by Chain Heal to targets beyond the first by 30%.
 		2: func(agent core.Agent) {
 			// Nothing to do
 		},
-		// After casting your Healing Wave, Lesser Healing Wave, or Riptide spell, gives you a 25% chance to gain Mana equal to 35% of the base cost of the spell.
+		// Improves your chance to get a critical strike with Nature spells by 3%.
 		4: func(agent core.Agent) {
-			// Not implementing for now
+			shaman := agent.(ShamanAgent).GetShaman()
+			shaman.PseudoStats.SchoolBonusCritChance[stats.SchoolIndexNature] += 3 * core.SpellCritRatingPerCritChance
 		},
-		// Your Healing Wave will now jump to additional nearby targets. Each jump reduces the effectiveness of the heal by 80%, and the spell will jump to up to 2 additional targets.
+		// When you cast a Healing Wave or Lesser Healing Wave, there is a 25% chance the target also receives a free Lightning Shield that causes 50 Nature damage to attacker on hit.
 		6: func(agent core.Agent) {
-			// Not implementing for now
-		},
-	},
-})
-
-var ItemSetEarthfuryImpact = core.NewItemSet(core.ItemSet{
-	Name: "Earthfury Impact",
-	Bonuses: map[int32]core.ApplyEffect{
-		// The radius of your totems that affect friendly targets is increased to 40 yd.
-		2: func(agent core.Agent) {
 			// Nothing to do
-		},
-		// Increases your critical strike chance with spells and attacks by 2%.
-		4: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStats(stats.Stats{
-				stats.MeleeCrit: 2 * core.CritRatingPerCritChance,
-				stats.SpellCrit: 2 * core.CritRatingPerCritChance,
-			})
-		},
-		// Your Flurry talent grants an additional 10% increase to your attack speed.
-		6: func(agent core.Agent) {
-			shaman := agent.(ShamanAgent).GetShaman()
-			shaman.RegisterAura(core.Aura{
-				Label: "S03 - Item - T1 - Shaman - Enhancement 6P Bonus",
-				OnInit: func(aura *core.Aura, sim *core.Simulation) {
-					shaman.bonusFlurrySpeed += .10
-				},
-			})
-		},
-	},
-})
-
-var ItemSetEarthfuryResolve = core.NewItemSet(core.ItemSet{
-	Name: "Earthfury Resolve",
-	Bonuses: map[int32]core.ApplyEffect{
-		// Increases your attack speed by 30% for your next 3 swings after you parry, dodge, or block.
-		2: func(agent core.Agent) {
-			shaman := agent.(ShamanAgent).GetShaman()
-
-			flurryAura := shaman.makeFlurryAura(5)
-			// The consumption trigger may not exist if the Shaman doesn't talent into Flurry
-			shaman.makeFlurryConsumptionTrigger(flurryAura)
-
-			core.MakePermanent(shaman.GetOrRegisterAura(core.Aura{
-				Label: "S03 - Item - T1 - Shaman - Tank 2P Bonus",
-				OnSpellHitTaken: func(aura *core.Aura, sim *core.Simulation, spell *core.Spell, result *core.SpellResult) {
-					if result.DidParry() || result.DidDodge() || result.DidBlock() {
-						flurryAura.Activate(sim)
-						flurryAura.SetStacks(sim, 3)
-					}
-				},
-			}))
-		},
-		// Your parries and dodges also activate your Shield Mastery rune ability.
-		4: func(agent core.Agent) {
-			// Implemented in runes.go
-		},
-		// Your Stoneskin Totem also reduces Physical damage taken by 5% and your Windwall Totem also reduces Magical damage taken by 5%.
-		6: func(agent core.Agent) {
-			// Debuffs implemented in core/buffs.go, activated with a raid buff setting or in earth_totems.go and air_totems.go
-		},
-	},
-})
-
-///////////////////////////////////////////////////////////////////////////
-//                            SoD Phase 5 Item Sets
-///////////////////////////////////////////////////////////////////////////
-
-var ItemSetEruptionOfTheTenStorms = core.NewItemSet(core.ItemSet{
-	Name: "Eruption of the Ten Storms",
-	Bonuses: map[int32]core.ApplyEffect{
-		// Your spell critical strikes now have a 100% chance trigger your Elemental Focus talent.
-		2: func(agent core.Agent) {
-			shaman := agent.(ShamanAgent).GetShaman()
-			if !shaman.Talents.ElementalFocus {
-				return
-			}
-
-			core.MakePermanent(shaman.RegisterAura(core.Aura{
-				Label: "S03 - Item - T2 - Shaman - Elemental 2P Bonus",
-				OnSpellHitDealt: func(aura *core.Aura, sim *core.Simulation, spell *core.Spell, result *core.SpellResult) {
-					if shaman.isShamanDamagingSpell(spell) && result.DidCrit() {
-						shaman.ClearcastingAura.Activate(sim)
-						shaman.ClearcastingAura.SetStacks(sim, shaman.ClearcastingAura.MaxStacks)
-					}
-				},
-			}))
-		},
-		// Loyal Beta from your Spirit of the Alpha ability now also increases Fire, Frost, and Nature damage by 5%.
-		4: func(agent core.Agent) {
-			shaman := agent.(ShamanAgent).GetShaman()
-			if !shaman.HasRune(proto.ShamanRune_RuneFeetSpiritOfTheAlpha) {
-				return
-			}
-
-			core.MakePermanent(shaman.RegisterAura(core.Aura{
-				Label: "S03 - Item - T2 - Shaman - Elemental 4P Bonus",
-				OnInit: func(aura *core.Aura, sim *core.Simulation) {
-					oldOnGain := shaman.LoyalBetaAura.OnGain
-					shaman.LoyalBetaAura.OnGain = func(aura *core.Aura, sim *core.Simulation) {
-						oldOnGain(aura, sim)
-						shaman.PseudoStats.SchoolDamageDealtMultiplier[stats.SchoolIndexFrost] *= 1.05
-						shaman.PseudoStats.SchoolDamageDealtMultiplier[stats.SchoolIndexFire] *= 1.05
-						shaman.PseudoStats.SchoolDamageDealtMultiplier[stats.SchoolIndexNature] *= 1.05
-					}
-				},
-			}))
-		},
-		// While Clearcasting is active, you deal 15% more non-Physical damage.
-		6: func(agent core.Agent) {
-			shaman := agent.(ShamanAgent).GetShaman()
-			if !shaman.Talents.ElementalFocus {
-				return
-			}
-
-			core.MakePermanent(shaman.RegisterAura(core.Aura{
-				Label: "S03 - Item - T2 - Shaman - Elemental 6P Bonus",
-				OnInit: func(aura *core.Aura, sim *core.Simulation) {
-					oldOnGain := shaman.ClearcastingAura.OnGain
-					shaman.ClearcastingAura.OnGain = func(aura *core.Aura, sim *core.Simulation) {
-						oldOnGain(aura, sim)
-						shaman.PseudoStats.SchoolDamageDealtMultiplier.MultiplyMagicSchools(1.15)
-					}
-
-					oldOnExpire := shaman.ClearcastingAura.OnExpire
-					shaman.ClearcastingAura.OnExpire = func(aura *core.Aura, sim *core.Simulation) {
-						oldOnExpire(aura, sim)
-						shaman.PseudoStats.SchoolDamageDealtMultiplier.MultiplyMagicSchools(1 / 1.15)
-					}
-				},
-			}))
 		},
 	},
 })
