+++ ui/index.html
@@ -180,25 +180,25 @@
 								</a>
 								<ul class="dropdown-menu w-100">
 									<li>
-										<a href="#" class="sim-link text-shaman">
+										<a href="/classic/elemental_shaman" class="sim-link text-shaman">
 											<div class="sim-link-content">
 												<img src="https://wow.zamimg.com/images/wow/icons/large/spell_nature_lightning.jpg" class="sim-link-icon" />
 												<div class="d-flex flex-column">
 													<span class="sim-link-label">Shaman</span>
 													<span class="sim-link-title">Elemental</span>
-													<span class="launch-status-label text-brand">Phase 1 - Unreleased</span>
+													<span class="launch-status-label text-brand">Phase 1 - Alpha</span>
 												</div>
 											</div>
 										</a>
 									</li>
 									<li>
-										<a href="#" class="sim-link text-shaman">
+										<a href="/classic/enhancement_shaman" class="sim-link text-shaman">
 											<div class="sim-link-content">
 												<img src="https://wow.zamimg.com/images/wow/icons/large/ability_shaman_stormstrike.jpg" class="sim-link-icon" />
 												<div class="d-flex flex-column">
 													<span class="sim-link-label">Shaman</span>
 													<span class="sim-link-title">Enhancement</span>
-													<span class="launch-status-label text-brand">Phase 1 - Unreleased</span>
+													<span class="launch-status-label text-brand">Phase 1 - Alpha</span>
 												</div>
 											</div>
 										</a>
